+++ x/metadata/client/cli/tx.go
@@ -9,6 +9,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/tx"
+	"github.com/cosmos/cosmos-sdk/version"
 
 	"github.com/provenance-io/provenance/x/metadata/types"
 
@@ -32,11 +33,17 @@ func NewTxCmd() *cobra.Command {
 
 	txCmd.AddCommand(
 		AddMetadataScopeCmd(),
+		RemoveMetadataScopeCmd(),
+
 		AddOsLocatorCmd(),
 		RemoveOsLocatorCmd(),
 		ModifyOsLocatorCmd(),
+
 		AddContractSpecificationCmd(),
 		RemoveContractSpecificationCmd(),
+
+		AddRecordSpecificationCmd(),
+		RemoveRecordSpecificationCmd(),
 	)
 
 	return txCmd
@@ -46,7 +53,7 @@ func NewTxCmd() *cobra.Command {
 func AddMetadataScopeCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "add-scope [scope-id] [spec-id] [owner-addresses] [data-access] [value-owner-address]",
-		Short: "Add a metadata scope to the provenance blockchain",
+		Short: "Add/Update a metadata scope to the provenance blockchain",
 		Args:  cobra.ExactArgs(5),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientTxContext(cmd)
@@ -82,13 +89,9 @@ func AddMetadataScopeCmd() *cobra.Command {
 			dataAccess := strings.Split(args[3], ",")
 			valueOwnerAddress := args[4]
 
-			signers := parseSigners(cmd, &clientCtx)
-			for _, signer := range signers {
-				_, err := sdk.AccAddressFromBech32(signer)
-				if err != nil {
-					fmt.Printf("signer address must be a Bech32 string: %v", err)
-					return err
-				}
+			signers, err := parseSigners(cmd, &clientCtx)
+			if err != nil {
+				return err
 			}
 
 			scope := *types.NewScope(
@@ -130,15 +133,13 @@ func RemoveMetadataScopeCmd() *cobra.Command {
 			scopeID, err = types.MetadataAddressFromBech32(args[0])
 			if err != nil {
 				return err
-			}
-			signers := parseSigners(cmd, &clientCtx)
 
-			for _, signer := range signers {
-				_, err := sdk.AccAddressFromBech32(signer)
-				if err != nil {
-					fmt.Printf("signer address must be a Bech32 string: %v", err)
-					return err
-				}
+			}
+			scopeMetaAddress := types.ScopeMetadataAddress(scopeUUID)
+
+			signers, err := parseSigners(cmd, &clientCtx)
+			if err != nil {
+				return err
 			}
 
 			deleteScope := *types.NewMsgDeleteScopeRequest(scopeID, signers)
@@ -257,13 +258,12 @@ func ModifyOsLocatorCmd() *cobra.Command {
 // AddContractSpecificationCmd creates a command to add/update contract specifications
 func AddContractSpecificationCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "add-contract-specification [contractspec-id] [owners] [parties-involved] [source-type] [source-value] [classname] [description-name] [description] [website-url] [icon-url]",
+		Use:   "add-contract-specification [contractspec-id] [owners] [parties-involved] [source-value] [classname] [description-name] [description] [website-url] [icon-url]",
 		Short: "Add/Update metadata contract specification on the provenance blockchain",
 		Long: `Add/Update metadata contract specification on the provenance blockchain
-[contractspec-id] - metaaddress of contract specification
+[contractspec-id] - contract specification metaaddress
 [owners] - comma delimited list of bech32 owner addresses
 [parties-involved] - comma delimited list of party types.  Accepted values: originator,servicer,investor,custodian,owner,affiliate,omnibus,provenance
-[source-type] - accepted values: hash or resourceid
 [source-value] - source identifier of type hash or resourceid
 [classname] - Name of contract specification
 [description-name] - optional- description name identifier 
@@ -271,58 +271,117 @@ func AddContractSpecificationCmd() *cobra.Command {
 [website-url] - optional - address of website
 [icon-url] - optional - address to a image to be used as an icon
 		`,
-		Args: cobra.RangeArgs(6, 10),
+		Args: cobra.RangeArgs(5, 9),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err
 			}
-
 			var specificationID types.MetadataAddress
 			specificationID, err = types.MetadataAddressFromBech32(args[0])
 			if err != nil {
 				return err
 			}
 
-			if !specificationID.IsContractSpecificationAddress() {
-				return fmt.Errorf("invalid contract specification id: %s", args[0])
-			}
-
 			partiesInvolved := parsePartyTypes(args[2])
-			description := parseDescription(args[6:])
+			description := parseDescription(args[5:])
 			contractSpecification := types.ContractSpecification{SpecificationId: specificationID,
 				Description:     description,
 				OwnerAddresses:  strings.Split(args[1], ","),
 				PartiesInvolved: partiesInvolved,
-				ClassName:       args[5],
+				ClassName:       args[4],
 			}
-			switch s := strings.ToUpper(args[3]); s {
-			case "RESOURCEID":
-				var recordID types.MetadataAddress
-				recordID, err = types.MetadataAddressFromBech32(args[4])
-				if err != nil {
-					return err
+			sourceValue := args[3]
+			var recordID types.MetadataAddress
+			recordID, err = types.MetadataAddressFromBech32(sourceValue)
+			if err != nil {
+				contractSpecification.Source = &types.ContractSpecification_Hash{
+					Hash: sourceValue,
 				}
+			} else {
 				contractSpecification.Source = &types.ContractSpecification_ResourceId{
 					ResourceId: recordID,
 				}
-			case "HASH":
-				contractSpecification.Source = &types.ContractSpecification_Hash{
-					Hash: args[4],
-				}
-			default:
-				return fmt.Errorf("incorrect source type for contract specification: %s", s)
 			}
 
-			signers := parseSigners(cmd, &clientCtx)
+			signers, err := parseSigners(cmd, &clientCtx)
+			if err != nil {
+				return err
+			}
 
 			msg := types.NewMsgAddContractSpecificationRequest(contractSpecification, signers)
 			err = msg.ValidateBasic()
 			if err != nil {
 				return err
 			}
-
 			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+	addSignerFlagCmd(cmd)
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func AddRecordSpecificationCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "add-record-specification [specification-id] [name] [input-specifications] [type-name] [result-types] [responsible-parties]",
+		Short: "Add/Update metadata record specification to the provenance blockchain",
+		Long: fmt.Sprintf(`Add/Update metadata record specification to the provenance blockchain.
+[specification-id] - record specification metaaddress
+[name] - record name
+[input-specifications] - semi-colon delimited list of input specifications <name>,<type-name>,<source-value>
+[type-name] - contract specification type name
+[result-types] - result definition type.  Accepted values: proposed,record,record_list
+[responsible-parties] - comma delimited list of party types.  Accepted values: originator,servicer,investor,custodian,owner,affiliate,omnibus,provenance
+
+Example: 
+$ %s tx metadata recspec1qh... recordname inputname1,typename1,hashvalue;inputename2,typename2,<recordmetaaddress> record_list owner,originator
+
+`, version.AppName),
+		Args: cobra.ExactArgs(6),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			specificationID, err := types.MetadataAddressFromBech32(args[0])
+			if err != nil {
+				return err
+			}
+
+			recordName := args[1]
+
+			inputs, err := parseInputSpecification(args[2])
+			if err != nil {
+				return err
+			}
+
+			resultType := definitionType(args[4])
+			partyTypes := parsePartyTypes(args[5])
+			signers, err := parseSigners(cmd, &clientCtx)
+			if err != nil {
+				return err
+			}
+
+			recordSpecification := types.RecordSpecification{
+				SpecificationId:    specificationID,
+				Name:               recordName,
+				Inputs:             inputs,
+				TypeName:           args[3],
+				ResultType:         resultType,
+				ResponsibleParties: partyTypes,
+			}
+
+			msg := *types.NewMsgAddRecordSpecificationRequest(recordSpecification, signers)
+
+			err = msg.ValidateBasic()
+			if err != nil {
+				return err
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), &msg)
 		},
 	}
 
@@ -332,17 +391,49 @@ func AddContractSpecificationCmd() *cobra.Command {
 	return cmd
 }
 
+// parseInputSpecification converts cli delimited argument and converts it to InputSpecifications
+func parseInputSpecification(cliDelimitedValue string) ([]*types.InputSpecification, error) {
+	delimitedInputs := strings.Split(cliDelimitedValue, ";")
+	inputs := make([]*types.InputSpecification, len(delimitedInputs))
+	for i, delimitedInput := range delimitedInputs {
+		values := strings.Split(delimitedInput, ",")
+		if len(values) != 3 {
+			return nil, fmt.Errorf("invalid number of values for input specification: %v", len(values))
+		}
+		inputs[i] = &types.InputSpecification{
+			Name:     values[0],
+			TypeName: values[1],
+		}
+		sourceValue := values[2]
+		recordID, err := types.MetadataAddressFromBech32(sourceValue)
+		if err != nil {
+			inputs[i].Source = &types.InputSpecification_Hash{Hash: sourceValue}
+		} else {
+			inputs[i].Source = &types.InputSpecification_RecordId{RecordId: recordID}
+		}
+	}
+	return inputs, nil
+}
+
 func addSignerFlagCmd(cmd *cobra.Command) {
 	cmd.Flags().String(FlagSigners, "", "comma delimited list of bech32 addresses")
 }
 
-func parseSigners(cmd *cobra.Command, client *client.Context) []string {
+func parseSigners(cmd *cobra.Command, client *client.Context) ([]string, error) {
 	flagSet := cmd.Flags()
 	if flagSet.Changed(FlagSigners) {
 		signerList, _ := flagSet.GetString(FlagSigners)
-		return strings.Split(signerList, ",")
+		signers := strings.Split(signerList, ",")
+		for _, signer := range signers {
+			_, err := sdk.AccAddressFromBech32(signer)
+			if err != nil {
+				fmt.Printf("signer address must be a Bech32 string: %v", err)
+				return nil, err
+			}
+		}
+		return signers, nil
 	}
-	return []string{client.GetFromAddress().String()}
+	return []string{client.GetFromAddress().String()}, nil
 }
 
 func parsePartyTypes(delimitedPartyTypes string) []types.PartyType {
@@ -353,6 +444,11 @@ func parsePartyTypes(delimitedPartyTypes string) []types.PartyType {
 		partyTypes[i] = types.PartyType(partyValue)
 	}
 	return partyTypes
+}
+
+func definitionType(cliValue string) types.DefinitionType {
+	typeValue := types.DefinitionType_value[fmt.Sprintf("DEFINITION_TYPE_%s", strings.ToUpper(cliValue))]
+	return types.DefinitionType(typeValue)
 }
 
 func parseDescription(cliArgs []string) *types.Description {
@@ -394,17 +490,53 @@ func RemoveContractSpecificationCmd() *cobra.Command {
 				return err
 			}
 
-			if !specificationID.IsContractSpecificationAddress() {
-				return fmt.Errorf("invalid contract specification id: %s", args[0])
+			signers, err := parseSigners(cmd, &clientCtx)
+			if err != nil {
+				return err
 			}
 
-			signers := parseSigners(cmd, &clientCtx)
 			msg := types.NewMsgDeleteContractSpecificationRequest(specificationID, signers)
 			err = msg.ValidateBasic()
 			if err != nil {
 				return err
 			}
 			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	addSignerFlagCmd(cmd)
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func RemoveRecordSpecificationCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "remove-record-specification [specification-id]",
+		Short: "Remove record specification from the provenance blockchain",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			var specificationID types.MetadataAddress
+			specificationID, err = types.MetadataAddressFromBech32(args[0])
+			if err != nil {
+				return err
+			}
+
+			if !specificationID.IsRecordSpecificationAddress() {
+				return fmt.Errorf("invalid contract specification id: %s", args[0])
+			}
+			signers, err := parseSigners(cmd, &clientCtx)
+			if err != nil {
+				return err
+			}
+			msg := *types.NewMsgDeleteRecordSpecificationRequest(specificationID, signers)
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), &msg)
 		},
 	}