+++ CHANGELOG.md
@@ -37,11 +37,15 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Bug Fixes
 
 * Resolves an issue where Gov Proposals to Create a new name would fail for new root domains #192
+* Remove deprecated ModuleCdc amino encoding from Metadata Locator records #187
 
 ### Features
 
 * Allow withdrawals of any coin type from a marker account in WASM smart contracts. #151
 * Added cli tx commands `add-contract-specification` `remove-contract-specification` for updating/adding/removing metadata `ContractSpecification`s. #195
+* Added cli tx commands `add-record-specification` `remove-record-specification` for updating/adding/removing metadata `RecordSpecification`s. #176
+* Simulation testing support #95
+* Name module simulation testing #24
 
 ## [v0.3.0](https://github.com/provenance-io/provenance/releases/tag/v0.3.0) - 2021-03-19